Added a flag to suppress mark all as read warning alert
